+++ project/app/services/pokemon_service.py
@@ -4,8 +4,8 @@ import random
 import pokebase as pb
 from dotenv import load_dotenv
 
-from logger import Logger
-from models import QuizInfo
+from util.logger import Logger
+from QuizInfo import QuizInfo
 
 load_dotenv()
 pb.cache.set_cache(os.getenv("POKEMON_CACHE"))
